+++ lib/StabilizedVehicle/src/MotorMixerBase.h
@@ -21,7 +21,7 @@ public:
 
     virtual void outputToMotors(const commands_t& commands, float deltaT, uint32_t tickCount) = 0;
     virtual float getMotorOutput(size_t motorIndex) const = 0;
-    virtual float getMotorRPM(size_t motorIndex) const = 0;
+    virtual int32_t getMotorRPM(size_t motorIndex) const = 0;
 public:
     static float clip(float value, float min, float max) { return value < min ? min : value > max ? max : value; }
 protected:
